+++ CredentialFormatter/DatabaseCredentialFormatter.php
@@ -1,0 +1,46 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Bartacus Platform.sh bundle.
+ *
+ * Copyright (c) Emily Karisch
+ *
+ * This bundle is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This bundle is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this bundle. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+namespace Bartacus\Bundle\PlatformshBundle\CredentialFormatter;
+
+class DatabaseCredentialFormatter
+{
+    public static function formatMysql(array $credentials): array
+    {
+        return array_merge([
+            'driver' => 'mysqli',
+            'charset' => 'utf8mb4',
+        ], self::formatCredentials($credentials));
+    }
+
+    private static function formatCredentials(array $credentials): array
+    {
+        return [
+            'host' => $credentials['host'],
+            'port' => $credentials['port'],
+            'dbname' => $credentials['path'],
+            'user' => $credentials['username'],
+            'password' => $credentials['password'],
+        ];
+    }
+}